This is the initial commit of spectralign
The object-oriented interface is new.

The underlying functional interface, src/spectralign/funcs.py,
is a direct copy of the python/swiftir.py from my earlier swift-ir
package that used to be hosted at
<https://gitlab.3dem.org/tmbartol/swift-ir.git>.

That python code, in turn, was a reimplementation of some of the
functionality of Art Wetzel’s original C implementation of
the SWIFT-IR algorithm.

A descendent of that C implementation currently lives at
<https://github.com/mcellteam/swift-ir>.